Added `<IdeaLaunchPad/>` component
