+++ packages/commonwealth/client/scripts/views/pages/Communities/Communities.tsx
@@ -1,7 +1,6 @@
 import { ExtendedCommunity } from '@hicommonwealth/schemas';
 import { ChainBase, ChainNetwork } from '@hicommonwealth/shared';
 import { findDenominationString } from 'helpers/findDenomination';
-import { useFlag } from 'hooks/useFlag';
 import React, { Fragment, useMemo, useRef, useState } from 'react';
 import { Link } from 'react-router-dom';
 import { Virtuoso } from 'react-virtuoso';
@@ -14,7 +13,6 @@ import { z } from 'zod';
 import { useFetchTokenUsdRateQuery } from '../../../state/api/communityStake/index';
 import { trpc } from '../../../utils/trpcClient';
 import { NewCommunityCard } from '../../components/CommunityCard';
-import LaunchIdeaCard from '../../components/LaunchIdeaCard';
 import { CWText } from '../../components/component_kit/cw_text';
 import { CWButton } from '../../components/component_kit/new_designs/CWButton';
 import CWCircleMultiplySpinner from '../../components/component_kit/new_designs/CWCircleMultiplySpinner';
@@ -26,7 +24,7 @@ import ManageCommunityStakeModal from '../../modals/ManageCommunityStakeModal/Ma
 import './Communities.scss';
 import { FiltersDrawer } from './FiltersDrawer/FiltersDrawer';
 import { CommunityFilters } from './FiltersDrawer/types';
-import TokenLaunchDrawer from './TokenLaunchDrawer';
+import IdeaLaunchpad from './IdeaLaunchpad';
 import { getCommunityCountsString } from './helpers';
 
 type ExtendedCommunityType = z.infer<typeof ExtendedCommunity>;
@@ -37,8 +35,6 @@ type ExtendedCommunitySliceType = [
 
 const CommunitiesPage = () => {
   const containerRef = useRef();
-
-  const tokenizedCommunityEnabled = useFlag('tokenizedCommunity');
 
   const {
     setModeOfManageCommunityStakeModal,
@@ -89,7 +85,6 @@ const CommunitiesPage = () => {
   const ethUsdRate = ethUsdRateData?.data?.data?.amount;
 
   const [isFilterDrawerOpen, setIsFilterDrawerOpen] = useState(false);
-  const [isTokenLaunchDrawerOpen, setIsTokenLaunchDrawerOpen] = useState(false);
 
   const isLoading =
     isLoadingTags ||
@@ -203,17 +198,7 @@ const CommunitiesPage = () => {
               onFiltersChange={(newFilters) => setFilters(newFilters)}
             />
           </div>
-          {tokenizedCommunityEnabled && (
-            <>
-              <LaunchIdeaCard
-                onTokenLaunchClick={() => setIsTokenLaunchDrawerOpen(true)}
-              />
-              <TokenLaunchDrawer
-                isOpen={isTokenLaunchDrawerOpen}
-                onClose={() => setIsTokenLaunchDrawerOpen(false)}
-              />
-            </>
-          )}
+          <IdeaLaunchpad />
         </div>
         {isLoading && communitiesList.length === 0 ? (
           <CWCircleMultiplySpinner />
