+++ packages/commonwealth/client/scripts/views/pages/Communities/IdeaLaunchpad/TokenLaunchDrawer/TokenLaunchDrawer.scss
@@ -1,4 +1,4 @@
-@import '../../../../../styles/shared.scss';
+@import '../../../../../../styles/shared.scss';
 
 .TokenLaunchDrawer {
   .Drawer.filter-drawer {
