+++ packages/commonwealth/client/scripts/views/pages/Communities/IdeaLaunchpad/index.ts
@@ -1,0 +1,3 @@
+import IdeaLaunchpad from './IdeaLaunchpad';
+
+export default IdeaLaunchpad;
